Adds exhaustive documentation to the main function about how modules interact with each other and what their are responsible for
